refactor(Navigation-Bar): move the nav component in a separate file.
