--- app/styles/home.css
@@ -1,25 +1,0 @@
-.primary_nav {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-  width: 96px;
-  min-height: 100vh;
-  height: 100%;
-  background: var(--semi-dark-blue);
-  border-radius: 20px;
-  margin: 32px;
-  padding: 3rem;
-}
-
-.navigation_list {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-evenly;
-  padding: 0;
-  margin: 0;
-  list-style-type: none;
-  height: 200px;
-  width: 100%;
-}
